Add files via upload
In this project, we are using two ultrasonic sensors. In this sketch, the code of one ultrasonic sensor with the buzzer is given. Using a serial monitor we will be able to test the distance up to which we want the garbage to be full. After that, the buzzer will make a sound.