Add files via upload
As I am planning to modify an Arduino based smart Dustbin. The actual project is done by the admin of the website Techatronics.com. In the original project, the dustbin has the ability to open the lid on its own when the ultrasonic sensor attached to it, senses presence like a human or any other object within a certain distance. This code is for opening the lid with the help of an ultrasonic sensor and servo.